requirements.txt
alchemy
pandas
psycopg2
sqlalchemy-redshift